+++ src/tr1/game/option.h
@@ -2,5 +2,6 @@
 
 #include "global/types.h"
 
-void Option_DoInventory(INVENTORY_ITEM *inv_item);
+void Option_Control(INVENTORY_ITEM *inv_item);
+void Option_Draw(INVENTORY_ITEM *inv_item);
 void Option_Shutdown(INVENTORY_ITEM *inv_item);