+++ src/tr1/game/option/option_compass.h
@@ -2,7 +2,7 @@
 
 #include "global/types.h"
 
-void Option_Compass(INVENTORY_ITEM *inv_item);
+void Option_Compass_Control(INVENTORY_ITEM *inv_item);
 void Option_Compass_Shutdown(void);
 
 void Option_Compass_UpdateNeedle(const INVENTORY_ITEM *inv_item);