+++ src/tr1/game/option/option_control_pick.h
@@ -1,6 +1,0 @@
-#pragma once
-
-#include "game/option/option_control.h"
-
-CONTROL_MODE Option_ControlPick(void);
-void Option_ControlPick_Shutdown(void);