+++ src/tr1/game/option/option_controls_pick.h
@@ -1,0 +1,6 @@
+#pragma once
+
+#include "game/option/option_controls.h"
+
+CONTROL_MODE Option_ControlsPick_Control(void);
+void Option_ControlsPick_Shutdown(void);