+++ src/tr1/game/option/option_graphics.h
@@ -2,5 +2,5 @@
 
 #include "global/types.h"
 
-void Option_Graphics(INVENTORY_ITEM *inv_item);
+void Option_Graphics_Control(INVENTORY_ITEM *inv_item);
 void Option_Graphics_Shutdown(void);