+++ src/tr1/game/option/option_passport.h
@@ -2,5 +2,5 @@
 
 #include "global/types.h"
 
-void Option_Passport(INVENTORY_ITEM *inv_item);
+void Option_Passport_Control(INVENTORY_ITEM *inv_item);
 void Option_Passport_Shutdown(void);